LayoutTests/accessibility/image-link-expected.txt

Image link in the presence of inline continuations

This test checks that the right accessibility tree is generated for a link inside an image


AXRole: AXLink
AXSubrole: (null)
AXRoleDescription: link
AXChildren: <array of size 1>
AXHelp:
AXParent: <AXLink: 'Delicious cake'>
AXSize: NSSize: {280, 215}
AXTitle: Delicious cake
AXDescription:
AXValue:
AXFocused: 1
AXEnabled: 1
AXWindow: <AXLink: 'Delicious cake'>
AXSelectedTextMarkerRange: (null)
AXStartTextMarker: <AXLink: 'Delicious cake'>
AXEndTextMarker: <AXLink: 'Delicious cake'>
AXVisited: 0
AXLinkedUIElements: <array of size 0>
AXSelected: 0
AXBlockQuoteLevel: 0
AXTopLevelUIElement: <AXLink: 'Delicious cake'>
AXLanguage:
AXDOMIdentifier: test
AXDOMClassList: <array of size 0>
AXFocusableAncestor: <AXLink: 'Delicious cake'>
AXEditableAncestor: (null)
AXHighestEditableAncestor: (null)
AXURL: http://www.wowhead.com/?item=33924
AXAccessKey: (null)
AXLinkRelationshipType:
AXElementBusy: 0


Child 0:
AXRole: AXImage
AXSubrole: (null)
AXRoleDescription: image
AXChildren: <array of size 0>
AXHelp:
AXParent: <AXImage>
AXSize: NSSize: {280, 210}
AXTitle:
AXDescription: Delicious cake
AXValue:
AXFocused: 0
AXEnabled: 1
AXWindow: <AXImage>
AXSelectedTextMarkerRange: (null)
AXStartTextMarker: <AXImage>
AXEndTextMarker: <AXImage>
AXVisited: 0
AXLinkedUIElements: <array of size 0>
AXSelected: 0
AXBlockQuoteLevel: 0
AXTopLevelUIElement: <AXImage>
AXLanguage:
AXDOMIdentifier:
AXDOMClassList: <array of size 0>
AXFocusableAncestor: <AXImage>
AXEditableAncestor: (null)
AXHighestEditableAncestor: (null)
AXURL: LayoutTests/accessibility/resources/cake.png
AXAccessKey: (null)
AXLinkRelationshipType:
AXElementBusy: 0
